Latent/x/backend/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.1
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
auditok==0.1.5
beautifulsoup4==4.8.0
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
cinemagoer==2023.5.1
click==8.1.8
colorama==0.4.1
comm==0.2.2
contourpy==1.2.0
cvlib==0.2.7
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
EbookLib==0.17.1
executing==2.2.0
faust-cchardet==2.1.19
ffmpeg==1.4
ffmpeg-python==0.2.0
ffsubsync==0.4.26
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.1.21
fonttools==4.48.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==2.8
imageio==2.37.0
imageio-ffmpeg==0.6.0
IMDbPY==2022.7.9
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
mira_sdk==0.1.14
ml-dtypes==0.4.1
mne==1.9.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
proglog==0.1.10
progressbar==2.5
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.1
PySocks==1.7.1
pysubs2==1.7.3
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==3.141.0
semantic-version==2.10.0
sentence-transformers==3.4.0
sentencepiece==0.2.0
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.14.0
SQLAlchemy==2.0.37
srt==3.5.3
stack-data==0.6.3
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
timm==1.0.14
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.48.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.67
ultralytics-thop==2.0.14
urllib3==1.26.6
wcwidth==0.2.13
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
